+++ packages/playground/vite.config.ts
@@ -2,16 +2,11 @@ import { defineConfig } from 'vite';
 import react from '@vitejs/plugin-react';
 import path from 'node:path';
 import { fileURLToPath } from 'url';
-import { hmrPlugins } from './scripts/hmrPlugin';
-
-const customLitPath = path.resolve(
-  __dirname,
-  '../blocks/src/__internal__/index.js'
-);
+import { hmrPlugin } from './scripts/hmr-plugin';
 
 // https://vitejs.dev/config/
 export default defineConfig({
-  plugins: [react(), hmrPlugins],
+  plugins: [react(), hmrPlugin],
   resolve: {
     alias: {
       '@blocksuite/blocks': path.resolve(
